requirement.txt
fastapi 
uvicorn 
asyncpg
pydantic-settings
sqlmodel
greenlet
alembic
passlib
bcrypt
python-jose
redis
fastapi-mail
itsdangerous
pytest
httpx